requirements.txt
Pillow==7.1.2
PyPDF4==1.27.0
Pypubsub==4.0.3
